core: Provide shared default implementations for RendererSealed's text functions
The skia, femtovg, anyrender, and Qt renderers all implemented
text_size, text_content_widths, char_size, font_metrics,
text_input_byte_offset_for_position,
text_input_cursor_rect_for_byte_offset, and the two font registration
functions with byte-identical delegation to the shared parley layout.

Give RendererSealed a text_layout_cache() accessor and, when the
shared-parley feature is enabled, default implementations of these
functions that route through it. Renderers now only override the
accessor; without the feature the functions remain required, as before.

The sharedparley entry points become generic over
(impl RendererSealed + ?Sized) so that the trait's default bodies can
pass self without unsized coercion.

The testing backend keeps its fixed-test-font overrides but drops its
identical font registration functions. The anyrender renderer keeps an
explicit None-returning text_content_widths override for now to
preserve its current layout behavior; removing it is a separate change
that updates the wrapped-text goldens. anyrender's
register_font_from_memory now registers static fonts without copying,
like the other renderers.
