+++ internal/core/renderer.rs
@@ -49,9 +49,19 @@ impl<T: RendererSealed> Renderer for T {}
 /// trait is not exported in the public API, it is not possible for the
 /// users to re-implement these functions.
 pub trait RendererSealed {
+    /// The cache used by the default implementations of the text related trait functions,
+    /// which lay text out through [`crate::textlayout::sharedparley`]. Renderers using
+    /// those default implementations return their cache here; with the default `None`,
+    /// text layouts are computed without caching.
+    #[cfg(feature = "shared-parley")]
+    fn text_layout_cache(&self) -> Option<&crate::textlayout::sharedparley::TextLayoutCache> {
+        None
+    }
+
     /// Returns the size of the given text in logical pixels.
     /// When set, `max_width` means that one need to wrap the text, so it does not go further than that,
     /// using the wrapping type passed by `text_wrap`.
+    #[cfg(not(feature = "shared-parley"))]
     fn text_size(
         &self,
         text_item: Pin<&dyn crate::item_rendering::RenderString>,
@@ -59,6 +69,30 @@ pub trait RendererSealed {
         max_width: Option<LogicalLength>,
         text_wrap: TextWrap,
     ) -> LogicalSize;
+
+    /// Returns the size of the given text in logical pixels.
+    /// When set, `max_width` means that one need to wrap the text, so it does not go further than that,
+    /// using the wrapping type passed by `text_wrap`.
+    ///
+    /// The default implementation uses the shared parley text layout with [`Self::text_layout_cache`].
+    #[cfg(feature = "shared-parley")]
+    fn text_size(
+        &self,
+        text_item: Pin<&dyn crate::item_rendering::RenderString>,
+        item_rc: &crate::item_tree::ItemRc,
+        max_width: Option<LogicalLength>,
+        text_wrap: TextWrap,
+    ) -> LogicalSize {
+        crate::textlayout::sharedparley::text_size(
+            self,
+            text_item,
+            item_rc,
+            max_width,
+            text_wrap,
+            self.text_layout_cache(),
+        )
+        .unwrap_or_default()
+    }
 
     /// Returns the content widths of the text, or None if the renderer can't measure them,
     /// in which case the caller falls back to `text_size` without a lower bound.
@@ -70,11 +104,19 @@ pub trait RendererSealed {
         text_item: Pin<&dyn crate::item_rendering::RenderString>,
         item_rc: &crate::item_tree::ItemRc,
     ) -> Option<ContentWidths> {
-        let _ = (text_item, item_rc);
-        None
+        #[cfg(feature = "shared-parley")]
+        {
+            crate::textlayout::sharedparley::text_content_widths(self, text_item, item_rc)
+        }
+        #[cfg(not(feature = "shared-parley"))]
+        {
+            let _ = (text_item, item_rc);
+            None
+        }
     }
 
     /// Returns the size of the individual character in logical pixels.
+    #[cfg(not(feature = "shared-parley"))]
     fn char_size(
         &self,
         text_item: Pin<&dyn crate::item_rendering::HasFont>,
@@ -82,14 +124,50 @@ pub trait RendererSealed {
         ch: char,
     ) -> LogicalSize;
 
+    /// Returns the size of the individual character in logical pixels.
+    ///
+    /// The default implementation measures the character with the shared parley layout.
+    #[cfg(feature = "shared-parley")]
+    fn char_size(
+        &self,
+        text_item: Pin<&dyn crate::item_rendering::HasFont>,
+        item_rc: &crate::item_tree::ItemRc,
+        ch: char,
+    ) -> LogicalSize {
+        self.slint_context()
+            .and_then(|ctx| {
+                let mut font_ctx = ctx.font_context().borrow_mut();
+                crate::textlayout::sharedparley::char_size(&mut font_ctx, text_item, item_rc, ch)
+            })
+            .unwrap_or_default()
+    }
+
     /// Returns the metrics of the given font.
+    #[cfg(not(feature = "shared-parley"))]
     fn font_metrics(&self, font_request: crate::graphics::FontRequest)
     -> crate::items::FontMetrics;
+
+    /// Returns the metrics of the given font.
+    ///
+    /// The default implementation queries the font through the shared fontique collection.
+    #[cfg(feature = "shared-parley")]
+    fn font_metrics(
+        &self,
+        font_request: crate::graphics::FontRequest,
+    ) -> crate::items::FontMetrics {
+        self.slint_context()
+            .map(|ctx| {
+                let mut font_ctx = ctx.font_context().borrow_mut();
+                crate::textlayout::sharedparley::font_metrics(&mut font_ctx, font_request)
+            })
+            .unwrap_or_default()
+    }
 
     /// Returns the (UTF-8) byte offset in the text property that refers to the character that contributed to
     /// the glyph cluster that's visually nearest to the given coordinate. This is used for hit-testing,
     /// for example when receiving a mouse click into a text field. Then this function returns the "cursor"
     /// position.
+    #[cfg(not(feature = "shared-parley"))]
     fn text_input_byte_offset_for_position(
         &self,
         text_input: Pin<&crate::items::TextInput>,
@@ -97,15 +175,59 @@ pub trait RendererSealed {
         pos: LogicalPoint,
     ) -> usize;
 
+    /// Returns the (UTF-8) byte offset in the text property that refers to the character that contributed to
+    /// the glyph cluster that's visually nearest to the given coordinate. This is used for hit-testing,
+    /// for example when receiving a mouse click into a text field. Then this function returns the "cursor"
+    /// position.
+    ///
+    /// The default implementation uses the shared parley text layout with [`Self::text_layout_cache`].
+    #[cfg(feature = "shared-parley")]
+    fn text_input_byte_offset_for_position(
+        &self,
+        text_input: Pin<&crate::items::TextInput>,
+        item_rc: &ItemRc,
+        pos: LogicalPoint,
+    ) -> usize {
+        crate::textlayout::sharedparley::text_input_byte_offset_for_position(
+            self,
+            text_input,
+            item_rc,
+            pos,
+            self.text_layout_cache(),
+        )
+    }
+
     /// That's the opposite of [`Self::text_input_byte_offset_for_position`]
     /// It takes a (UTF-8) byte offset in the text property, and returns a Rectangle
     /// left to the char. It is one logical pixel wide and ends at the baseline.
+    #[cfg(not(feature = "shared-parley"))]
     fn text_input_cursor_rect_for_byte_offset(
         &self,
         text_input: Pin<&crate::items::TextInput>,
         item_rc: &ItemRc,
         byte_offset: usize,
     ) -> LogicalRect;
+
+    /// That's the opposite of [`Self::text_input_byte_offset_for_position`]
+    /// It takes a (UTF-8) byte offset in the text property, and returns a Rectangle
+    /// left to the char. It is one logical pixel wide and ends at the baseline.
+    ///
+    /// The default implementation uses the shared parley text layout with [`Self::text_layout_cache`].
+    #[cfg(feature = "shared-parley")]
+    fn text_input_cursor_rect_for_byte_offset(
+        &self,
+        text_input: Pin<&crate::items::TextInput>,
+        item_rc: &ItemRc,
+        byte_offset: usize,
+    ) -> LogicalRect {
+        crate::textlayout::sharedparley::text_input_cursor_rect_for_byte_offset(
+            self,
+            text_input,
+            item_rc,
+            byte_offset,
+            self.text_layout_cache(),
+        )
+    }
 
     /// Clear the caches for the items that are being removed
     fn free_graphics_resources(
@@ -121,7 +243,7 @@ pub trait RendererSealed {
     /// Example: when a PopupWindow disappears, the region under the popup needs to be redrawn
     fn mark_dirty_region(&self, _region: crate::partial_renderer::DirtyRegion) {}
 
-    #[cfg(feature = "std")] // FIXME: just because of the Error
+    #[cfg(all(feature = "std", not(feature = "shared-parley")))] // FIXME: just because of the Error
     /// This function can be used to register a custom TrueType font with Slint,
     /// for use with the `font-family` property. The provided slice must be a valid TrueType
     /// font.
@@ -132,7 +254,22 @@ pub trait RendererSealed {
         Err("This renderer does not support registering custom fonts.".into())
     }
 
-    #[cfg(feature = "std")]
+    #[cfg(all(feature = "std", feature = "shared-parley"))]
+    /// This function can be used to register a custom TrueType font with Slint,
+    /// for use with the `font-family` property. The provided slice must be a valid TrueType
+    /// font.
+    ///
+    /// The default implementation registers the font with the shared fontique collection.
+    fn register_font_from_memory(
+        &self,
+        data: &'static [u8],
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        let ctx = self.slint_context().ok_or("slint platform not initialized")?;
+        ctx.font_context().borrow_mut().register_static_font(data);
+        Ok(())
+    }
+
+    #[cfg(all(feature = "std", not(feature = "shared-parley")))]
     /// This function can be used to register a custom TrueType font with Slint,
     /// for use with the `font-family` property. The provided path must refer to a valid TrueType
     /// font.
@@ -141,6 +278,23 @@ pub trait RendererSealed {
         _path: &std::path::Path,
     ) -> Result<(), Box<dyn std::error::Error>> {
         Err("This renderer does not support registering custom fonts.".into())
+    }
+
+    #[cfg(all(feature = "std", feature = "shared-parley"))]
+    /// This function can be used to register a custom TrueType font with Slint,
+    /// for use with the `font-family` property. The provided path must refer to a valid TrueType
+    /// font.
+    ///
+    /// The default implementation registers the font with the shared fontique collection.
+    fn register_font_from_path(
+        &self,
+        path: &std::path::Path,
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        let requested_path = path.canonicalize().unwrap_or_else(|_| path.into());
+        let contents = std::fs::read(requested_path)?;
+        let ctx = self.slint_context().ok_or("slint platform not initialized")?;
+        ctx.font_context().borrow_mut().collection.register_fonts(contents.into(), None);
+        Ok(())
     }
 
     fn register_bitmap_font(&self, _font_data: &'static crate::graphics::BitmapFont) {
